hw6/pyproject.toml
[tool.poetry]
name = "hw6"
version = "0.1.0"
description = ""
authors = ["Eric860730 <s12345d60@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21"
scipy = "^1.6.1"
matplotlib = "^3.4"
pyqt5 = "^5.15"
libsvm-official = "^3.25"
pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"
rich = "^10.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
